% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chores.R
\docType{data}
\name{chores}
\alias{chores}
\title{Evaluation results}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
chores
}
\description{
The \code{chores} data contains "confirmed" evaluation results generated by
running the following:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tsk <- chores_task()
tsk$eval(
  solver_chat = ellmer::chat_*(model = "some-model")
)
}\if{html}{\out{</div>}}

See \code{\link[=chores_task]{chores_task()}} for more on how the evaluation task works. Notably:
\itemize{
\item The solver carries out 34 refactorings using
the \link[chores:`doc-helper-cli`]{cli chore helper}, each repeated 3 times.
\item Each refactoring is then graded according to a rubric using Claude 4
Sonnet. The grading results in a score between 0 and 1 and incorporates
measures of code quality as well as execution time.
\item Grading costs something like $2.50; the cost of solving depends on the
model pricing.
}
}
\section{Columns}{

\itemize{
\item \code{provider}:
\item \code{model}:
\item \code{score}:
\item \code{metadata}:
}
}

\keyword{datasets}
