% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{chores_dataset}
\alias{chores_dataset}
\title{The chores eval}
\format{
A tibble with 34 rows and 5 columns:
\describe{
\item{id}{Character. Unique identifier/title for the code problem.}
\item{input}{The user prompt that would be submitted to the chore.}
\item{target}{Character. The solution, often with a description of notable
features of a correct solution.}
\item{source}{Character. URL or source of the problem, usually as a link to
a commit on GitHub. \code{NA}s indicate that
the problem was written originally for this eval.}
}
}
\usage{
chores_dataset
}
\description{
Pass this dataset to \code{Task$new()} to situate it inside of an evaluation
task.
}
\examples{
chores_dataset

str(chores_dataset)
}
\seealso{
\code{\link[=chores_task]{chores_task()}} to situate this dataset in the chores eval.
}
\keyword{datasets}
