% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R
\name{chores_scorer}
\alias{chores_scorer}
\title{The chores scorer}
\usage{
chores_scorer(
  samples,
  ...,
  scorer_chat = ellmer::chat_anthropic(model = "claude-3-7-sonnet-latest")
)
}
\arguments{
\item{samples}{The samples from a solver task, likely retrieved from
\code{task$get_samples()}.}

\item{...}{Additional arguments passed to the scoring function.}

\item{scorer_chat}{An ellmer chat object to use for scoring. Defaults to
\code{ellmer::chat_anthropic(model = "claude-3-7-sonnet-latest")}; this is the
scoring model used in "official" results.}
}
\value{
A list with the following components:
\describe{
\item{score}{Numeric vector of scores between 0 and 1, representing the
proportion of criteria met.}
\item{.scorer_metadata}{List containing the prompts used for scoring and
the detailed grading results.}
}
}
\description{
Pass this function to \code{Task$new()} as the scorer to evaluate model responses
based on the chores eval criteria.
}
\seealso{
\link{chores_dataset} for the dataset this scorer evaluates, and
\code{\link[=chores_task]{chores_task()}} to combine this scorer with the dataset and solver.
}
