% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{chores_solver}
\alias{chores_solver}
\title{The chores solver}
\usage{
chores_solver(inputs, ..., solver_chat)
}
\arguments{
\item{inputs}{Character vector of user prompts, likely from
\code{chores_dataset$input}.}

\item{...}{Additional arguments passed to the \code{chat_parallel} method.}

\item{solver_chat}{An ellmer chat object to use for solving the prompts.
This must be a Chat object from the ellmer package.}
}
\value{
A list with the following components:
\describe{
\item{result}{Character vector of model responses, one for each input.}
\item{solver_chat}{List of Chat objects used to generate each response,
with the same length as \code{inputs}.}
}
}
\description{
Pass this function to \code{Task$new()} as the solver to process input prompts
from the chores dataset with a specified language model.
}
\seealso{
\link{chores_dataset} for the dataset this solver processes, and
\code{\link[=chores_task]{chores_task()}} to combine this solver with the chores dataset and scorer.
}
